requirements.txt
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.7
idna==3.7
pycparser==2.22
pycryptodome==3.20.0
requests==2.31.0
requests-pkcs12==1.24
urllib3==2.2.1
